fix: check bridge always return success
Due to the [bug](https://github.com/leapdao/tendermint/pull/10) in our tendermint fork, `chainHeight` is always 0
in a period handler. We didn't use it though for some reason and took block height from `chainInfo` instead. However,
it was changed in April and we missed the fact that `chainHeight` is always 0, thus making checkBridge ABCI to always
succeed without checking anything.