fix: relay events from Bridge just once.
For many Plasma Bridge events we are creating and publishing corresponding txs to mutate the app state.
This is done in `EventRelay` instance.

When the node is restarted in resumes operating (listening for Plasma Bridge events) from the previous period height, not from the last observed block.
E.g. if stopped at height 36, the node will replay blocks 32-36 upon restart and will resume listening for Bridge events from block 32. This leads
to the case when some of the events may be read one more time on restart and relayed one more time creating duplicate transactions which is not desirable. Consider the example
above any bridge events happened on the root chain after the 32 plasma block was processed will be read one more time on restart.

Proposed solution is to make EventRelay to "remember" the root chain height it was processed last time and ignore any events before that height.